+++ Attribute/Runtime/HexNumberAttribute.cs
@@ -1,0 +1,13 @@
+﻿
+using UnityEngine;
+
+namespace Hananoki {
+	/// <summary>
+	/// intに付けて16進数で表示する
+	/// 編集できないので注意
+	/// </summary>
+	public class HexNumberAttribute : PropertyAttribute {
+		public HexNumberAttribute() { }
+	}
+}
+