+++ RuntimeEditor/Package.cs
@@ -5,7 +5,7 @@ namespace Hananoki.SharedModule {
   public static class Package {
     public const string name = "SharedModule";
     public const string editorPrefName = "Hananoki.SharedModule";
-    public const string version = "1.5.2";
+    public const string version = "1.5.3";
   }
   
 #if UNITY_EDITOR