+++ RuntimeEditor/UnityEditorMenu.cs
@@ -3,6 +3,12 @@
 namespace Hananoki {
 	public static partial class UnityEditorMenu {
 
+		public static void Window_Hananoki_Settings() {
+			if( EditorHelper.IsDefine( "ENABLE_HANANOKI_SETTINGS" ) ) {
+				EditorApplication.ExecuteMenuItem( "Window/Hananoki/Settings" );
+			}
+		}
+
 		public static void File_Build_Settings() {
 			EditorApplication.ExecuteMenuItem( "File/Build Settings..." );
 		}